* Associating block numbers to events in the credit module; * Adding the ability to obtain the remaining earnings eras for the campaign attended; * Refactoring the credit module to take the source code written in lib.rs and abstract it into different files depending on the function.
