+++ pallets/credit/src/mock.rs
@@ -25,6 +25,7 @@ use sp_core::H256;
 use sp_runtime::{
     testing::Header,
     traits::{BlakeTwo256, IdentityLookup},
+    Percent,
 };
 
 use node_primitives::{Balance, BlockNumber, Moment};
